+++ src/clint/clint.sv
@@ -21,12 +21,12 @@ module clint #(
     parameter int unsigned AXI_DATA_WIDTH = 64,
     parameter int unsigned AXI_ID_WIDTH   = 10,
     parameter int unsigned NR_CORES       = 1 // Number of cores therefore also the number of timecmp registers and timer interrupts
-)(
+) (
     input  logic                clk_i,       // Clock
     input  logic                rst_ni,      // Asynchronous reset active low
     input  logic                testmode_i,
-    AXI_BUS.Slave               slave,
-
+    input  ariane_axi::req_t    axi_req_i,
+    output ariane_axi::resp_t   axi_resp_o,
     input  logic                rtc_i,       // Real-time clock in (usually 32.768 kHz)
     output logic [NR_CORES-1:0] timer_irq_o, // Timer interrupts
     output logic [NR_CORES-1:0] ipi_o        // software interrupt (a.k.a inter-process-interrupt)
@@ -35,6 +35,9 @@ module clint #(
     localparam logic [15:0] MSIP_BASE     = 16'h0;
     localparam logic [15:0] MTIMECMP_BASE = 16'h4000;
     localparam logic [15:0] MTIME_BASE    = 16'hbff8;
+
+    localparam AddrSelWidth = (NR_CORES == 1) ? 1 : $clog2(NR_CORES);
+
     // signals from AXI 4 Lite
     logic [AXI_ADDR_WIDTH-1:0] address;
     logic                      en;
@@ -60,14 +63,15 @@ module clint #(
         .AXI_DATA_WIDTH ( AXI_DATA_WIDTH ),
         .AXI_ID_WIDTH   ( AXI_ID_WIDTH    )
     ) axi_lite_interface_i (
-        .clk_i     ( clk_i   ),
-        .rst_ni    ( rst_ni  ),
-        .slave     ( slave   ),
-        .address_o ( address ),
-        .en_o      ( en      ),
-        .we_o      ( we      ),
-        .data_i    ( rdata   ),
-        .data_o    ( wdata   )
+        .clk_i      ( clk_i      ),
+        .rst_ni     ( rst_ni     ),
+        .axi_req_i  ( axi_req_i  ),
+        .axi_resp_o ( axi_resp_o ),
+        .address_o  ( address    ),
+        .en_o       ( en         ),
+        .we_o       ( we         ),
+        .data_i     ( rdata      ),
+        .data_o     ( wdata      )
     );
 
     // -----------------------------
@@ -86,11 +90,11 @@ module clint #(
         if (en && we) begin
             case (register_address) inside
                 [MSIP_BASE:MSIP_BASE+8*NR_CORES]: begin
-                    msip_n[$unsigned(address[NR_CORES-1+3:3])] = wdata[0];
+                    msip_n[$unsigned(address[AddrSelWidth-1+3:3])] = wdata[0];
                 end
 
                 [MTIMECMP_BASE:MTIMECMP_BASE+8*NR_CORES]: begin
-                    mtimecmp_n[$unsigned(address[NR_CORES-1+3:3])] = wdata;
+                    mtimecmp_n[$unsigned(address[AddrSelWidth-1+3:3])] = wdata;
                 end
 
                 MTIME_BASE: begin
@@ -108,11 +112,11 @@ module clint #(
         if (en && !we) begin
             case (register_address) inside
                 [MSIP_BASE:MSIP_BASE+8*NR_CORES]: begin
-                    rdata = msip_q[$unsigned(address[NR_CORES-1+3:3])];
+                    rdata = msip_q[$unsigned(address[AddrSelWidth-1+3:3])];
                 end
 
                 [MTIMECMP_BASE:MTIMECMP_BASE+8*NR_CORES]: begin
-                    rdata = mtimecmp_q[$unsigned(address[NR_CORES-1+3:3])];
+                    rdata = mtimecmp_q[$unsigned(address[AddrSelWidth-1+3:3])];
                 end
 
                 MTIME_BASE: begin