+++ src/plic/plic.sv
@@ -28,6 +28,8 @@
 //-------------------------------------------------------------------------------
 
 module plic #(
+    parameter int ADDR_WIDTH         = 32,   // can be either 32 or 64 bits (don't use 64bit at the moment as this causes memory map issues)
+    parameter int DATA_WIDTH         = 32,   // can be either 32 or 64 bits (don't use 64bit at the moment as this causes memory map issues)
     parameter int ID_BITWIDTH        = -1,   // width of the gateway identifiers
     parameter int PARAMETER_BITWIDTH = -1,   // width of the internal parameter e.g. priorities
     parameter int NUM_TARGETS        = -1,   // number of target slices
@@ -39,8 +41,6 @@ module plic #(
     output logic [NUM_TARGETS-1:0] eip_targets_o,
     REG_BUS.in                     external_bus_io
 );
-    localparam int ADDR_WIDTH         = 32;
-    localparam int DATA_WIDTH         = 32;
     // declare all local variables
     // gateway arrays always go from NUM_SOURCES to 1 because gateway ids start at 1
     logic                          gateway_irq_pendings    [NUM_SOURCES];               //for pending irqs of the gateways